Meals page feature fe (#29)
* fixed getting the meals logged by the user on a specific date, fixed getting the food informations from the database and updated its routes

* updated the theme of the frontend of food tracking page,
linked it to the backend for getting the meals, deleting the meals on date. did not make searching and adding meals.
updated the UI for only one meal.

* added leading when getting meals, adding navigation between dates

* A new coloring constants.
added the feature of searching for food and adding it to a date for tracking calories.
added some error handling
improved ui and added close buttons
added fetching api functions in lib folder

* fixing notch problem ( adding safe are view ) updating the url of the requests

* updated the theme of the frontend of food tracking page,
linked it to the backend for getting the meals, deleting the meals on date. did not make searching and adding meals.
updated the UI for only one meal.

* A new coloring constants.
added the feature of searching for food and adding it to a date for tracking calories.
added some error handling
improved ui and added close buttons
added fetching api functions in lib folder

* fixing notch problem ( adding safe are view ) updating the url of the requests

---------

Co-authored-by: Djilali Benahmed <dbenahmedd@gmail.com>